agency_platform/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-ratelimit>=4.1,<5.0
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0
